refactor(server): isolate local operation relay
